LPD-58977 Remove the unnecessary QueryFilterTranslator abstraction
